Fix ESLint errors in scroll memory and index files
- Convert consecutive line comments to block comments in scrollMemory.js
- Capitalize comment first letters
- Use object destructuring for pathname
- Replace ternary with if/else
- Add error logging to catch blocks
- Convert negated condition to positive in index.js
- Extract initFilterModal, initFontChanger, initImageLozad to separate modules
- Reduce index.js from 344 to 298 lines to meet max-lines rule
