+++ website/modules/asset/ui/src/initImageLozad.js
@@ -1,0 +1,8 @@
+import lozad from 'lozad';
+
+function initImageLozad() {
+  const observer = lozad();
+  observer.observe();
+}
+
+export { initImageLozad };
